+++ LPsolver.h
@@ -8,6 +8,6 @@
 
 # include "gurobi_interface.h"
 
-int guess_aux(game *gptr, float threshold,GRBenv *env);
+int guess_aux(game *gptr, double threshold,GRBenv *env);
 
 int guess_hint_aux(game *gptr, int row, int col,GRBenv *env);
